Implementacion promedio y personaje vs Liga

+++ src/Heroe.java
@@ -1,7 +1,9 @@
 public class Heroe extends Competidor {
 
-	public Heroe(String nombreReal, String nombrePersonaje, int velocidad, int fuerza, int resistencia, int destreza) {
+	public Heroe(String nombreReal, String nombrePersonaje, double velocidad, double fuerza, double resistencia,
+			double destreza) {
 		super(nombreReal, nombrePersonaje, velocidad, fuerza, resistencia, destreza);
+
 	}
 
 }
